+++ src/devices/IrrigationSystem.ts
@@ -78,23 +78,35 @@ export class IrrigationSystem extends DeviceBase {
       })
       .onSet(this.setActive.bind(this));
 
-    this.irrigation.service.getCharacteristic(this.platform.Characteristic.ProgramMode).onGet(() => {
-      return this.irrigation.service.getCharacteristic(this.platform.Characteristic.ProgramMode).value;
-    });
+    this.irrigation.service
+      .getCharacteristic(this.platform.Characteristic.ProgramMode)
+      .onGet(() => {
+        return this.irrigation.service.getCharacteristic(this.platform.Characteristic.ProgramMode).value;
+      });
 
-    this.irrigation.service.getCharacteristic(this.platform.Characteristic.InUse).onGet(() => {
-      this.rainbird!.refreshStatus();
-      return this.irrigation.InUse;
-    });
+    this.irrigation.service
+      .getCharacteristic(this.platform.Characteristic.InUse)
+      .onGet(() => {
+        this.rainbird!.refreshStatus();
+        return this.irrigation.InUse;
+      });
 
-    this.irrigation.service.getCharacteristic(this.platform.Characteristic.StatusFault).onGet(() => {
-      return this.irrigation.service.getCharacteristic(this.platform.Characteristic.StatusFault).value;
-    });
+    this.irrigation.service
+      .getCharacteristic(this.platform.Characteristic.StatusFault)
+      .onGet(() => {
+        return this.irrigation.service.getCharacteristic(this.platform.Characteristic.StatusFault).value;
+      });
 
-    this.irrigation.service.getCharacteristic(this.platform.Characteristic.RemainingDuration).onGet(() => {
-      this.rainbird!.refreshStatus();
-      return this.rainbird!.RemainingDuration();
-    });
+    this.irrigation.service
+      .getCharacteristic(this.platform.Characteristic.RemainingDuration)
+      .setProps({
+        minValue: device.minValueRemainingDuration,
+        maxValue: device.maxValueRemainingDuration! * rainbird!.zones.length,
+      })
+      .onGet(() => {
+        this.rainbird!.refreshStatus();
+        return this.rainbird!.remainingDuration();
+      });
 
     // Valves for zones
     for (const zone of rainbird!.zones) {
@@ -182,9 +194,13 @@ export class IrrigationSystem extends DeviceBase {
       this.valves
         .get(zone)!
         .service.getCharacteristic(this.platform.Characteristic.RemainingDuration)
+        .setProps({
+          minValue: device.minValueRemainingDuration,
+          maxValue: device.maxValueRemainingDuration,
+        })
         .onGet(() => {
           this.rainbird!.refreshStatus();
-          return this.rainbird!.RemainingDuration(zone);
+          return this.rainbird!.remainingDuration(zone);
         });
     }
 
@@ -264,12 +280,12 @@ export class IrrigationSystem extends DeviceBase {
       this.irrigation.service.updateCharacteristic(this.platform.Characteristic.InUse, this.irrigation.InUse);
       this.debugLog(`${this.constructor.name} ${this.accessory.displayName} updateCharacteristic InUse: ${this.irrigation.InUse}`);
     }
-    if (this.rainbird!.RemainingDuration() === undefined) {
-      this.debugLog(`${this.constructor.name} ${this.accessory.displayName} RemainingDuration: ${this.rainbird!.RemainingDuration()}`);
+    if (this.rainbird!.remainingDuration() === undefined) {
+      this.debugLog(`${this.constructor.name} ${this.accessory.displayName} RemainingDuration: ${this.rainbird!.remainingDuration()}`);
     } else {
-      this.irrigation.service.updateCharacteristic(this.platform.Characteristic.RemainingDuration, this.rainbird!.RemainingDuration());
+      this.irrigation.service.updateCharacteristic(this.platform.Characteristic.RemainingDuration, this.rainbird!.remainingDuration());
       this.debugLog(
-        `${this.constructor.name} ${this.accessory.displayName} updateCharacteristic RemainingDuration: ${this.rainbird!.RemainingDuration()}`);
+        `${this.constructor.name} ${this.accessory.displayName} updateCharacteristic RemainingDuration: ${this.rainbird!.remainingDuration()}`);
     }
 
     // Valves
@@ -298,11 +314,11 @@ export class IrrigationSystem extends DeviceBase {
         valve.service.updateCharacteristic(this.platform.Characteristic.IsConfigured, this.accessory.context.configured[zone]);
         this.debugLog(`${this.constructor.name} Valve ${zone} updateCharacteristic IsConfigured: ${this.accessory.context.configured[zone]}`);
       }
-      if (this.rainbird!.RemainingDuration(zone) === undefined) {
-        this.debugLog(`${this.constructor.name} Valve ${zone} RemainingDuration: ${this.rainbird!.RemainingDuration(zone)}`);
+      if (this.rainbird!.remainingDuration(zone) === undefined) {
+        this.debugLog(`${this.constructor.name} Valve ${zone} RemainingDuration: ${this.rainbird!.remainingDuration(zone)}`);
       } else {
-        valve.service.updateCharacteristic(this.platform.Characteristic.RemainingDuration, this.rainbird!.RemainingDuration(zone));
-        this.debugLog(`${this.constructor.name} Valve ${zone} updateCharacteristic RemainingDuration: ${this.rainbird!.RemainingDuration(zone)}`);
+        valve.service.updateCharacteristic(this.platform.Characteristic.RemainingDuration, this.rainbird!.remainingDuration(zone));
+        this.debugLog(`${this.constructor.name} Valve ${zone} updateCharacteristic RemainingDuration: ${this.rainbird!.remainingDuration(zone)}`);
       }
     }
   }