+++ src/settings.ts
@@ -12,7 +12,6 @@ export const PLUGIN_NAME = 'homebridge-rainbird';
 //Config
 export interface RainbirdPlatformConfig extends PlatformConfig {
   devices?: Array<DevicesConfig>;
-  disablePlugin?: boolean;
   options?: options | Record<string, never>;
 }
 
@@ -26,12 +25,19 @@ export type DevicesConfig = {
   showProgramCSwitch?: boolean;
   showProgramDSwitch?: boolean;
   showStopIrrigationSwitch?: boolean;
+  minValueRemainingDuration?: number;
+  maxValueRemainingDuration?: number;
   syncTime?: boolean;
   showRequestResponse?: boolean;
   showZoneValve?: boolean;
+  includeZones?: string,
+  showDelayIrrigationSwitch?: boolean;
+  irrigationDelay?: number;
+  external: boolean;
   firmware?: number;
   refreshRate?: number;
   logging?: string;
+  delete?: boolean;
 };
 
 export type options = {